Conduct Bayesian Calibration on An EnergyPlus Model

bayes_job() takes an IDF and EPW as input and returns a BayesCalibJob, which provides a prototype of conducting Bayesian calibration of EnergyPlus model.

Usage

bc <- bayes_job(idf, epw)
bc$version()
bc$seed()
bc$weather()
bc$read_rdd(update = FALSE)
bc$read_mdd(update = FALSE)
bc$input(key_value = NULL, name = NULL, reporting_frequency = NULL, append = FALSE)
bc$output(key_value = NULL, name = NULL, reporting_frequency = NULL, append = FALSE)
bc$param(..., .names = NULL, .num_sim = 30L)
bc$apply_measure(measure, ..., .num_sim = 30L)
bc$samples()
bc$models()
bc$eplus_run(dir = NULL, run_period = NULL, wait = TRUE, force = FALSE,
             copy_external = FALSE, echo = wait)
bc$data_sim(resolution = NULL, exclude_ddy = TRUE, all = FALSE)
bc$data_field(output, new_input = NULL, all = FALSE)
bc$data_bc(data_field = NULL, data_sim = NULL)
bc$stan_run(file = NULL, data = NULL, iter = 2000L, chains = 4L, echo = TRUE,
            mc.cores = parallel::detectCores(), ...)
bc$stan_file(path = NULL)
bc$eplus_kill()
bc$eplus_status()
bc$eplus_output_dir(which = NULL)
bc$eplus_locate_output(which = NULL, suffix = ".err", strict = TRUE)
bc$eplus_errors(which = NULL, info = FALSE)
bc$eplus_report_data_dict(which = NULL)
bc$eplus_report_data(which = NULL, key_value = NULL, name = NULL,
                     year = NULL, tz = "UTC", all = FALSE, wide = FALSE,
                     period = NULL, month = NULL, day = NULL, hour = NULL, minute = NULL,
                     interval = NULL, simulation_days = NULL, day_type = NULL,
                     environment_name = NULL)
bc$eplus_tabular_data(which = NULL, report_name = NULL, report_for = NULL,
                      table_name = NULL, column_name = NULL, row_name = NULL)
bc$eplus_save(dir = NULL, separate = TRUE, copy_external = FALSE)
bc$print()

Create

bc <- bayes_job(idf, epw)

When calling bayes_job(), the objects of classes related in output variable reporting will be deleted, in order to make sure all input and output variable specifications can be achieved using Output:Variable and Output:Meter. Classes to be deleted include:

Output:Variable

Output:Meter

Output:Meter:MeterFileOnly

Output:Meter:Cumulative

Output:Meter:Cumulative:MeterFileOnly

Meter:Custom

Meter:CustomDecrement

Output:EnvironmentalImpactFactors

Arguments

idf: Path to EnergyPlus IDF file or an Idf object.

epw: Path to EnergyPlus EPW file or an Epw object.

Get Seed Model and Weather

param$version()
param$seed()
param$weather()

$version() returns the version of input Idf object.

$seed() returns the input Idf object.

$weather() returns the input Epw object.

Get RDD amd MDD Data of Seed Model

param$read_rdd(update = FALSE)
param$read_mdd(update = FALSE)

$read_rdd() and $read_mdd() silently runs EnergyPlus using input seed model with design-day-only mode to create .rdd and .mdd file and returns the corresponding RddFile and MddFile object, respectively. The RddFile and MddFile object is stored internally and will be directly returned whenever you call $read_rdd() and $read_mdd() again. You can force to run the design-day-only simulation again to update the contents by setting update to TRUE.

$read_rdd() and read_mdd() is useful when adding input and output parameters.

Arguments

update: Whether to run the design-day-only simulation and parse .rdd and .mdd file again. Default: FALSE.

Set Input and Output Parameters

bc$input(key_value = NULL, name = NULL, reporting_frequency = NULL, append = FALSE)
bc$output(key_value = NULL, name = NULL, reporting_frequency = NULL, append = FALSE)

$input() and $output() takes input and output parameter definitions in a similar pattern as you set output variables in Output:Variable and Output:Meter class and returns a data.table containing the information of input and output parameters. For $input(), only variables in RDD are allowed. For $output(), both variables in RDD and MDD are allowd. The returned data.table has 5 columns:

index: Indices of input or output parameters.

class: The class that parameters belong to. Will be either Output:Variable or Output:Meter.

key_value: Key value name for variables.

variable_name: Variable names listed in RDD or MDD.

reporting_frequency: Variable reporting frequency.

If calling without any argument, the existing input and output parameters are directly returned, e.g. bc$input() and bc$output().

You can remove all existing input and output parameter by setting append to NULL, e.g. bc$input(append = NULL) and bc$output(append = NULL).

key_value accepts 3 different formats:

A character vector.

An RddFile or an MddFile object. They can be retrieved using $read_rdd() and $read_mdd(). In this case, name argument will be ignored, as its values are directly taken from variable names in input RddFile and MddFile object. For example:

bc$input(bc$read_rdd()[1:5])
bc$output(bc$read_mdd()[1:5])

A data.frame() with valid format for adding Output:Variable and Output:Meter objects using Idf$load(). In this case, name argument will be ignored. For example:

bc$input(eplusr::rdd_to_load(bc$read_rdd()[1:5]))
bc$output(eplusr::mdd_to_load(bc$read_mdd()[1:5]))

Arguments

key_value: Key value name for variables. If not specified, "*" are used for all variables. key_value can also be an RddFile, MddFile or a data.frame(). Please see description above.

name: Variable names listed in RDD or MDD.

reporting_frequency: Variable reporting frequency for all variables. If NULL, "Timestep" are used for all variables. All possible values: "Detailed", "Timestep", "Hourly", "Daily", "Monthly", "RunPeriod", "Environment", and "Annual". Default: NULL.

append: Whether to append input variables at the end of existing ones. A special value NULL can be given to remove all existing parameters. Default: FALSE.

Set Parameters

bc$param(..., .names = NULL, .num_sim = 30L)
bc$apply_measure(measure, ..., .num_sim = 30L)
bc$samples()
bc$models()

There are 2 ways to set calibration parameters in BayesCalibJob class, i.e. $param() and $apply_measure().

$param() takes parameter definitions in list format, which is similar to $set() in eplusr::Idf class except that each field is not assigned with a single value, but a numeric vector of length 2, indicating the minimum and maximum of the parameter. Every list in $param() should be named with a valid object name. Object ID can also be used but have to be combined with prevailing two periods .., e.g. ..10 indicates the object with ID 10.

There is two special syntax in $param():

class := list(field = value): Note the use of := instead of =. The main difference is that, unlike =, the left hand side of := should be a valid class name in the seed model. It will treat the specified field of all objects in specified class to as a single calibration parameter.

.(object, object) := list(field = value): Simimar like above, but note the use of .() in the left hand side. You can put multiple object ID or names in .(). It will treat the field of all specified objects to as a single calibration parameter.

For example, the code block below defines 4 calibration parameters:

Field Fan Total Efficiency in object named Supply Fan 1 in class Fan:VariableVolume class, with minimum and maximum being 0.1 and 1.0, respectively.

Field Thickness in all objects in class Material, with minimum and maximum being 0.01 and 1.0, respectively.

Field Conductivity in all objects in class Material, with minimum and maximum being 0.1 and 0.6, respectively.

Field Watts per Zone Floor Area in objects Light1 and Light2 in class Lights, with minimum and maximum being 10 and 30, respectively.

bc$param(
    `Supply Fan 1` = list(Fan_Total_Efficiency = c(min = 0.1, max = 1.0)),
    Material := list(Thickness = c(0.01, 1), Conductivity = c(0.1, 0.6)),
   .("Light1", "Light2") := list(Watts_per_Zone_Floor_Area = c(10, 30))
)

Arguments

...: Lists of paramter definitions. Each list should be named with a valid object name or a valid object ID denoted in style ..1, ..2 and etc. For specifying the fields for all objects in a class, the class name instead of the object name should be used, and a special notation := should be used instead of =, e.g. class := list(field = value). For grouping fields from different objects in the same class, use .() in the left hand side and put object ID or names inside., .e.g .(object1, object2) := list(field = value).

.num_sim: An positive integer specifying the number of simulations to run for each combination of calibration parameter value. Default: 30L.

.names: A character vector of the parameter names. If NULL, the parameter will be named in format t + number. Default: NULL.

$apply_measure() works in a similar way as the $apply_measure in eplusr::ParametricJob class, with only exception that each argument supplied in ... should be a numeric vector of length 2, indicating the minimum and maximum of the calibration parameter. Basically $apply_measure() allows to apply a measure to an Idf. A measure here is just a function that takes an Idf object and other arguments as input, and returns a modified Idf object as output. The names of function parameter will be used as the names of calibration parameter. For example, the equivalent version of specifying parameters described above using $apply_measure() can be:

measure <- function (idf, efficiency, thickness, conducitivy, lpd) {
    idf$set(
        `Supply Fan 1` = list(Fan_Total_Efficiency = efficiency),
        Material := list(Thickness = thickness, Conductivity = conducivity)
        .("Light1", "Light2") := list(Watts_per_Zone_Floor_Area = lpd)
    )

    idf
}

bc$apply_measure(measure,
    efficiency = c(min = 0.1, max = 1.0),
    thickness = c(0.01, 1), conductivity = c(0.1, 0.6),
    lpd = c(10, 30)
)

Arguments

measure: A function that takes an Idf and other arguments as input and returns an Idf object as output.

...: Arguments except first Idf argument that are passed to that measure.

.num_sim: An positive integer specifying the number of simulations to run for each value of calibration parameter value. (NOT CORRECT). Default: 30L.

All models created using $param() and $apply_measure() will be named in the same pattern, i.e. Case_ParameterName(ParamterValue).... Note that only paramter names will be abbreviated using abbreviate() with minlength being 5L and use.classes being TRUE. If samples contain duplications, make.unique() will be called to make sure every model has a unique name.

$samples() returns a data.table which contains the sampled value for each parameter using Random Latin Hypercube Sampling method. The returned data.table has 1 + n columns, where n is the parameter number, and 1 indicates an extra column named case giving the index of each sample. Note that if $samples() is called before input and output parameters being set using $input() and $output(), only the sampling will be performed and no parametric models will be created. This is because information of input and output parameters are needed in order to make sure that corresponding variables will be reported during simulations. In this case, you can use $models() to create those models.

$models() returns a list of parametric Idf objects created using calibration parameter values genereated using Random Latin Hypercube Sampling. As stated above, parametric models can only be created after input, output and calibration parameters have all be set using $input(), $output() and $param() (or $apply_measure()), respectively.

All models will be named in the same pattern, i.e. Case_ParameterName(ParamterValue).... Note that paramter names will be abbreviated using abbreviate() with minlength being 5L and use.classes being TRUE.

Run Parametric Simulations

bc$eplus_run(dir = NULL, run_period = NULL, wait = TRUE, force = FALSE,
             copy_external = FALSE, echo = wait)

$eplus_run() runs all parametric models in parallel. Parameter run_period can be given to insert a new RunPeriod object. In this case, all existing RunPeriod objects in the seed model will be commented out.

Note that when run_period is given, value of field Run Simulation for Weather File Run Periods in SimulationControl class will be reset to Yes to make sure input run period can take effect.

Arguments

dir: The parent output directory for specified simulations. Outputs of each simulation are placed in a separate folder under the parent directory. If NULL, directory of seed model will be used. Default: NULL.

run_period: A list giving a new RunPeriod object definition. If not NULL, only this new RunPeriod will take effect with all existing RunPeriod objects in the seed model being commented out. If NULL, existing run period in the seed model will be used. Default: NULL.

wait: If TRUE, R will hang on and wait all EnergyPlus simulations finish. If FALSE, all EnergyPlus simulations are run in the background. Default: TRUE.

force: Only applicable when the last simulation runs with wait equals to FALSE and is still running. If TRUE, current running job is forced to stop and a new one will start. Default: FALSE.

copy_external: If TRUE, the external files that every Idf object depends on will also be copied into the simulation output directory. The values of file paths in the Idf will be changed automatically. Currently, only Schedule:File class is supported. This ensures that the output directory will have all files needed for the model to run. Default is FALSE.

echo: Only applicable when wait is TRUE. Whether to print simulation status. Default: the value of wait.

Collect Simulation Data

bc$data_sim(resolution = NULL, exclude_ddy = TRUE, all = FALSE)

$data_sim() returns a list of 2 data.table which contains the simulated data of input and output parameters. These data will be stored internally and used during Bayesian calibration using Stan.

The resolution parameter can be used to specify the time resolution of returned data. Note that input time resolution cannot be smaller than the reporting frequency, otherwise an error will be issued.

The parameter is named in the same way as standard EnergyPlus csv output file, i.e. KeyValue:VariableName[Unit](Frequency).

By default, $data_sim() returns minimal columns, i.e. the Date/Time column together with all input and output parameters are returned. You can retrieve extra columns by setting all to TRUE. Those column include:

case: Integer type. Indices of parametric simulations.

environment_period_index: Integer type. The indice of environment.

environment_name: Character type. A text string identifying the simulation environment.

simulation_days: Integer type. Day of simulation.

datetime: DateTime type. The date time of simulation result. Note that the year valueas are automatically calculated to meets the start day of week restriction for each simulation environment.

month: Integer type. The month of reported date time.

day: Integer type. The day of month of reported date time.

hour: Integer type. The hour of reported date time.

minute: Integer type. The minute of reported date time.

day_type: Character type. The type of day, e.g. Monday, Tuesday and etc. Note that day_type will always be NA if resolution is specified.

Arguments

resolution: A character string specifying a time unit or a multiple of a unit to change the time resolution of returned simulation data. Valid base units are min, hour, day, week, month, and year. Example: 10 mins, 2 hours, 1 day. If NULL, the variable reporting frequency is used. Default: NULL.

exclude_ddy: Whether to exclude design day data. Default: TRUE. Default: FALSE.

all: If TRUE, extra columns are also included in the returned data.table describing the simulation case and datetime components. Default: FALSE.

Specify Measured Data

bc$data_field(output, new_input = NULL, all = FALSE)

$data_field() takes a data.frame() of measured value of output parameters and returns a list of data.tables which contains the measured value of input and output parameters, and newly measured value of input if applicable.

The specified output data.frame() is validated using criteria below:

The column number should be the same as the number of output specified in $output().

The row number should be the same as the number of simulated values for each case extracted using $data_sim().

For input parameters, the values of simulation data for the first case are directly used as the measured values.

Parameter new_input can be used to give a data.frame() of newly measured value of input parameters. The column number of input data.frame() should be the same as the number of input parameters specified in $input(). If not specified, the measured values of input parameters will be used for predictions.

All the data will be stored internally and used during Bayesian calibration using Stan.

Note that as $data_field() relies on the output of $data_sim() to perform validation on the specified data, $data_field() cannot be called before $data_sim() and internally stored data will be removed whenever $data_sim() is called. This aims to make sure that simulated data and field data can be matched whenever the calibration is performed.

Arguments

output: A data.frame() containing measured value of output parameters.

new_input: A data.frame() containing newly measured value of input parameters.

all: If TRUE, extra columns are also included in the returned data.table describing the simulation case and datetime components. For details, please see $data_sim(). Default: FALSE.

Run Bayesian Calibration Using Stan

bc$data_bc(data_field = NULL, data_sim = NULL)
bc$stan_run(file = NULL, data = NULL, iter = 2000L, chains = 4L, echo = TRUE,
            mc.cores = parallel::detectCores(), ...)
bc$stan_file(path = NULL)

$data_bc() takes a list of field data and simulated data, and returns a list that contains data input for Bayesican calibration using the Stan model from Chong (2018):

n: Number of measured parameter observations.

n_pred: Number of newly design points for predictions.

m: Number of simulated observations.

p: Number of input parameters.

q: Number of calibration parameters.

yf: Data of measured output after z-score standardization using data of simulated output.

yc: Data of simulated output after z-score standardization.

xf: Data of measured input after min-max normalization.

xc: Data of simulated input after min-max normalization.

x_pred: Data of new design points for predictions after min-max normalization.

tc: Data of calibration parameters after min-max normalization.

Input data_field and data_sim should have the same structure as the output from $data_field() and $data_sim(). If data_field and data_sim is not specified, the output from $data_field() and $data_sim() will be used.

$stan_run() runs Bayesian calibration using Stan and returns a list of 2 elements:

fit: An object of S4 class rstan::stanfit.

y_pred: A data.table with predicted output values.

$stan_file() saves the Stan file used internally for Bayesican calibration. If no path is given, a character vector of the Stan code is returned. If given, the code will be save to the path and the file path is returned.

Arguments

file: The path to the Stan program to use. If NULL, the pre-compiled Stan code from Chong (2018) will be used. Default: NULL.

data: Only applicable when file is not NULL. The data to be used for Bayesican calibration. If NULL, the data that $data_bc() returns is used. Default: NULL.

path: A path to save the Stan code. If NULL, a character vector of the Stan code is returned.

iter: A positive integer specifying the number of iterations for each chain (including warmup). Default: 2000.

chains: A positive integer specifying the number of Markov chains. Default: 4.

echo: Whether to print intermediate output from Stan on the console, which might be helpful for model debugging. Default: TRUE.

mc.cores: An integer specifying how many cores to be used for Stan. Default: parallel::detectCores().

...: Additional arguments to pass to rstan::sampling (when file is NULL) or rstan::stan (when file is not NULL).

Inherited Methods from ParametricJob

bc$eplus_kill()
bc$eplus_status()
bc$eplus_output_dir(which = NULL)
bc$eplus_locate_output(which = NULL, suffix = ".err", strict = TRUE)
bc$eplus_errors(which = NULL, info = FALSE)
bc$eplus_report_data_dict(which = NULL)
bc$eplus_report_data(which = NULL, key_value = NULL, name = NULL,
                     year = NULL, tz = "UTC", all = FALSE, wide = FALSE,
                     period = NULL, month = NULL, day = NULL, hour = NULL, minute = NULL,
                     interval = NULL, simulation_days = NULL, day_type = NULL,
                     environment_name = NULL)
bc$eplus_tabular_data(which = NULL, report_name = NULL, report_for = NULL,
                      table_name = NULL, column_name = NULL, row_name = NULL)
bc$eplus_save(dir = NULL, separate = TRUE, copy_external = FALSE)

All methods listed above are inherited from eplusr's ParametricJob. Each method has been renamed with a prefix eplus_, e.g. $output_dir() in eplusr::ParametricJob becomes $eplus_output_dir(). For detailed documentation on each method, please see eplusr's documentation.

References

A. Chong and K. Menberg, "Guidelines for the Bayesian calibration of building energy models", Energy and Buildings, vol. 174, pp. 527–547. DOI: 10.1016/j.enbuild.2018.06.028

Examples

if (FALSE) {
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "5Zone_Transformer.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    bayes_job(idf_path, epw_path)

    # create from an Idf and an Epw object
    bc <- bayes_job(read_idf(idf_path), read_epw(epw_path))

    # get the seed model
    bc$seed()

    # get the weather
    bc$weather()

    # read rdd
    bc$read_rdd()

    # read mdd
    bc$read_mdd()

    # set input and output parameters
    bc$input(name = "fan air mass flow rate", reporting_frequency = "hourly")
    # set input parameters
    bc$output(name = "fan electric power", reporting_frequency = "hourly")

    # get existing input and output
    bc$input()
    bc$output()

    # set sensitivity parameters using $param()
    bc$param(`Supply Fan 1` = list(Fan_Total_Efficiency = c(min = 0.1, max = 1.0)),
        .names = "faneff", .num_sim = 2
    )

    # extract samples
    bc$samples()

    # extract all models
    bc$models()

    # run parametric simulations
    bc$eplus_run(dir = tempdir(), run_period = list("example", 1, 1, 1, 31))

    # print simulation errors
    bc$eplus_errors()

    # extract simulation data
    bc$data_sim()

    # specify field data
    # here use the seed model as an example
    ## clone the seed model
    seed <- bc$seed()$clone()
    ## remove existing RunPeriod objects
    seed$RunPeriod <- NULL
    ## set run period as the same as in `$eplus_run()`
    seed$add(RunPeriod = list("test", 1, 1, 1, 31))
    ## save the model to tempdir
    seed$save(tempfile(fileext = ".idf"))
    job <- seed$run(epw_path, echo = FALSE)
    fan_power <- job$report_data(name = bc$output()$variable_name, wide = TRUE)
    bc$data_field(fan_power[, -c("case", "Date/Time")])

    # run stan
    fit <- bc$stan_run()
}
}